+++ constant.py
@@ -1,7 +1,7 @@
 FRED_MD_SERIES = "https://api.stlouisfed.org/fred/series"
 FRED_MD_SERIES_CATEGORIES = "https://api.stlouisfed.org/fred/series/categories"
 FRED_MD_SERIES_OBSERVATIONS = "https://api.stlouisfed.org/fred/series/observations"
-
+FRED_MD_SERIES_SEARCH = "https://api.stlouisfed.org/fred/series/search"
 
 OBSERVATIONS_PARAM = [
     "realtime_start",